fix(earn): cast pgx params to int in LEAST() + show server error in UI
pgx sends parameters as text inside LEAST() because PostgreSQL cannot
infer types for polymorphic functions. Explicit ::int casts fix the
"column points is of type integer but expression is of type text" error.

Frontend now shows "잠시 후에 다시 시도해주세요" on server errors instead
of incorrectly displaying "획득 완료".

Constraint: pgx cannot infer parameter OIDs inside polymorphic SQL functions
Confidence: high
Scope-risk: narrow

+++ server/storage/level_repository.go
@@ -74,9 +74,9 @@ func (r *LevelRepository) EarnCoin(ctx context.Context, userID string, runID, co
 	var newTotal int
 	err = tx.QueryRow(ctx, `
 		INSERT INTO user_points (user_id, points, updated_at)
-		VALUES ($1, LEAST($2, $3), NOW())
+		VALUES ($1, LEAST($2::int, $3::int), NOW())
 		ON CONFLICT (user_id) DO UPDATE SET
-			points = LEAST(user_points.points + $2, $3),
+			points = LEAST(user_points.points + $2::int, $3::int),
 			updated_at = NOW()
 		RETURNING points
 	`, userID, coins, effectiveCap).Scan(&newTotal)
